+++ HealthKitReporter/Classes/Decorator/Extensions+HKWorkout.swift
@@ -12,13 +12,26 @@ extension HKWorkout: Harmonizable {
     typealias Harmonized = Workout.Harmonized
 
     func harmonize() throws -> Harmonized {
-        let totalEnergyBurnedUnit = HKUnit.largeCalorie()
-        let totalEnergyBurned = self.totalEnergyBurned?.doubleValue(for: totalEnergyBurnedUnit) ?? 0
+        let totalEnergyBurnedUnit: HKUnit
+        if #available(iOS 11.0, *) {
+            totalEnergyBurnedUnit = HKUnit.largeCalorie()
+        } else {
+            totalEnergyBurnedUnit = HKUnit.kilocalorie()
+        }
+        let totalEnergyBurned = self.totalEnergyBurned?.doubleValue(for: totalEnergyBurnedUnit)
+        
         let totalDistanceUnit = HKUnit.meter()
-        let totalDistance = self.totalDistance?.doubleValue(for: totalDistanceUnit) ?? 0
+        let totalDistance = self.totalDistance?.doubleValue(for: totalDistanceUnit)
+
         let countUnit = HKUnit.count()
-        let totalSwimmingStrokeCount = self.totalSwimmingStrokeCount?.doubleValue(for: countUnit) ?? 0
-        let totalFlightsClimbed = self.totalFlightsClimbed?.doubleValue(for: countUnit) ?? 0
+        var totalSwimmingStrokeCount: Double?
+        if #available(iOS 10.0, *) {
+            totalSwimmingStrokeCount = self.totalSwimmingStrokeCount?.doubleValue(for: countUnit)
+        }
+        var totalFlightsClimbed: Double?
+        if #available(iOS 11.0, *) {
+            totalFlightsClimbed = self.totalFlightsClimbed?.doubleValue(for: countUnit)
+        }
         return Harmonized(
             value: Int(workoutActivityType.rawValue),
             totalEnergyBurned: totalEnergyBurned,