+++ HealthKitReporter/Classes/Model/Payload/Category.swift
@@ -23,6 +23,18 @@ public struct Category: Identifiable, Sample {
             self.description = description
             self.metadata = metadata 
         }
+
+        public func copyWith(
+            value: Int? = nil,
+            description: String? = nil,
+            metadata: [String: String]? = nil
+        ) -> Harmonized {
+            return Harmonized(
+                value: value ?? self.value,
+                description: description ?? self.description,
+                metadata: metadata ?? self.metadata
+            )
+        }
     }
 
     public let uuid: String
@@ -77,6 +89,24 @@ public struct Category: Identifiable, Sample {
         self.device = device
         self.sourceRevision = sourceRevision
         self.harmonized = harmonized
+    }
+
+    public func copyWith(
+        identifier: String? = nil,
+        startTimestamp: Double? = nil,
+        endTimestamp: Double? = nil,
+        device: Device? = nil,
+        sourceRevision: SourceRevision? = nil,
+        harmonized: Harmonized? = nil
+    ) -> Category {
+        return Category(
+            identifier: identifier ?? self.identifier,
+            startTimestamp: startTimestamp ?? self.startTimestamp,
+            endTimestamp: endTimestamp ?? self.endTimestamp,
+            device: device ?? self.device,
+            sourceRevision: sourceRevision ?? self.sourceRevision,
+            harmonized: harmonized ?? self.harmonized
+        )
     }
 }
 // MARK: - Original