+++ HealthKitReporter/Classes/Model/Payload/Device.swift
@@ -48,6 +48,33 @@ public struct Device: Codable {
         self.localIdentifier = localIdentifier
         self.udiDeviceIdentifier = udiDeviceIdentifier
     }
+
+    public static func local() -> Device {
+        let local = HKDevice.local()
+        return Device(device: local)
+    }
+
+    public func copyWith(
+        name: String? = nil,
+        manufacturer: String? = nil,
+        model: String? = nil,
+        hardwareVersion: String? = nil,
+        firmwareVersion: String? = nil,
+        softwareVersion: String? = nil,
+        localIdentifier: String? = nil,
+        udiDeviceIdentifier: String? = nil
+    ) -> Device {
+        return Device(
+            name: name ?? self.name,
+            manufacturer: manufacturer ?? self.manufacturer,
+            model: model ?? self.model,
+            hardwareVersion: hardwareVersion ?? self.hardwareVersion,
+            firmwareVersion: firmwareVersion ?? self.firmwareVersion,
+            softwareVersion: softwareVersion ?? self.softwareVersion,
+            localIdentifier: localIdentifier ?? self.localIdentifier,
+            udiDeviceIdentifier: udiDeviceIdentifier ?? self.udiDeviceIdentifier
+        )
+    }
 }
 // MARK: - Original
 extension Device: Original {