+++ HealthKitReporter/Classes/Model/Payload/SourceRevision.swift
@@ -37,6 +37,18 @@ public struct SourceRevision: Codable {
             self.minorVersion = minorVersion
             self.patchVersion = patchVersion
         }
+
+        public func copyWith(
+            majorVersion: Int? = nil,
+            minorVersion: Int? = nil,
+            patchVersion: Int? = nil
+        ) -> OperatingSystem {
+            return OperatingSystem(
+                majorVersion: majorVersion ?? self.majorVersion,
+                minorVersion: minorVersion ?? self.minorVersion,
+                patchVersion: patchVersion ?? self.patchVersion
+            )
+        }
     }
 
     public let source: Source
@@ -48,11 +60,21 @@ public struct SourceRevision: Codable {
     init(sourceRevision: HKSourceRevision) {
         self.source = Source(source: sourceRevision.source)
         self.version = sourceRevision.version
-        self.productType = sourceRevision.productType
-        self.systemVersion = sourceRevision.systemVersion
-        self.operatingSystem = OperatingSystem(
-            version: sourceRevision.operatingSystemVersion
-        )
+        if #available(iOS 11.0, *) {
+            self.productType = sourceRevision.productType
+            self.systemVersion = sourceRevision.systemVersion
+            self.operatingSystem = OperatingSystem(
+                version: sourceRevision.operatingSystemVersion
+            )
+        } else {
+            self.productType = nil
+            self.systemVersion = "1.0.0"
+            self.operatingSystem = OperatingSystem(
+                majorVersion: 1,
+                minorVersion: 0,
+                patchVersion: 0
+            )
+        }
     }
 
     public init(
@@ -68,16 +90,38 @@ public struct SourceRevision: Codable {
         self.systemVersion = systemVersion
         self.operatingSystem = operatingSystem
     }
+
+    public func copyWith(
+        source: Source? = nil,
+        version: String? = nil,
+        productType: String? = nil,
+        systemVersion: String? = nil,
+        operatingSystem: OperatingSystem? = nil
+    ) -> SourceRevision {
+        return SourceRevision(
+            source: source ?? self.source,
+            version: version ?? self.version,
+            productType: productType ?? self.productType,
+            systemVersion: systemVersion ?? self.systemVersion,
+            operatingSystem: operatingSystem ?? self.operatingSystem
+        )
+    }
 }
 // MARK: - Original
 extension SourceRevision: Original {
     func asOriginal() throws -> HKSourceRevision {
-        return HKSourceRevision(
-            source: try source.asOriginal(),
-            version: version,
-            productType: productType,
-            operatingSystemVersion: operatingSystem.original
-        )
+        if #available(iOS 11.0, *) {
+            return HKSourceRevision(
+                source: try source.asOriginal(),
+                version: version,
+                productType: productType,
+                operatingSystemVersion: operatingSystem.original
+            )
+        } else {
+            throw HealthKitError.notAvailable(
+                "HKSourceRevision is not available for the current iOS"
+            )
+        }
     }
 }
 // MARK: - Payload