+++ HealthKitReporter/Classes/Model/Type/CharacteristicType.swift
@@ -30,7 +30,11 @@ public enum CharacteristicType: Int, CaseIterable, ObjectType {
         case .biologicalSex:
             return HKObjectType.characteristicType(forIdentifier: .biologicalSex)
         case .wheelchairUse:
-            return HKObjectType.characteristicType(forIdentifier: .wheelchairUse)
+            if #available(iOS 10.0, *) {
+                return HKObjectType.characteristicType(forIdentifier: .wheelchairUse)
+            } else {
+                return nil
+            }
         case .activityMoveMode:
             if #available(iOS 14.0, *) {
                 return HKObjectType.characteristicType(forIdentifier: .activityMoveMode)